backend/requirements.txt
python-dotenv==1.0.1
openai==1.68.2
fastapi==0.115.11
fastapi-cli==0.0.7
uvicorn==0.34.0
requests==2.32.3
yfinance==0.2.55
langgraph==0.3.20
langchain_core==0.3.48
langchain_openai==0.3.10